+++ install/create_console_objects.sql
@@ -15,7 +15,6 @@ prompt - Project page https://github.com/ogobrecht/console
 prompt - Set compiler flags
 declare
   v_apex_installed varchar2(5) := 'FALSE'; -- Do not change (is set dynamically).
-  v_apex_fun       varchar2(5) := 'FALSE'; -- Bring some fun into the APEX error handling ;-)
   v_utils_public   varchar2(5) := 'FALSE'; -- Make utilities public available (for testing or other usages).
 begin
 
@@ -34,7 +33,6 @@ begin
 
   execute immediate 'alter session set plsql_ccflags = '''
     || 'APEX_INSTALLED:' || v_apex_installed || ','
-    || 'APEX_FUN:      ' || v_apex_fun       || ','
     || 'UTILS_PUBLIC:'   || v_utils_public   || '''';
 
 end;
@@ -58,15 +56,15 @@ begin
         units_level_info     varchar2 (4000 byte)            ,
         units_level_debug    varchar2 (4000 byte)            ,
         units_level_trace    varchar2 (4000 byte)            ,
+        enable_ascii_art     varchar2 (   1 byte)  not null  ,
         --
-        constraint  console_conf_pk   primary key ( conf_id                 )  ,
-        constraint  console_conf_ck1  check       ( conf_id = 'GLOBAL_CONF' )  ,
-        --
-        constraint  console_conf_ck2  check ( level_id   in (1, 2, 3)                                          )  ,
-        constraint  console_conf_ck3  check ( level_name =  decode(level_id, 1,'error', 2,'warning', 3,'info') )  ,
-        --
-        constraint  console_conf_ck4  check ( check_interval between 10 and 60 )
-      ) organization index
+        constraint  console_conf_pk   primary key ( conf_id )                                                    ,
+        constraint  console_conf_ck1  check ( conf_id = 'GLOBAL_CONF' )                                          ,
+        constraint  console_conf_ck2  check ( level_id in (1, 2, 3) )                                            ,
+        constraint  console_conf_ck3  check ( level_name = decode(level_id, 1,'error', 2,'warning', 3,'info') )  ,
+        constraint  console_conf_ck4  check ( check_interval between 10 and 60 )                                 ,
+        constraint  console_conf_ck5  check ( enable_ascii_art in ('Y','N') )
+      )
     }';
   else
     dbms_output.put_line('- Table CONSOLE_CONF found, no action required');
@@ -86,6 +84,7 @@ comment on column console_conf.units_level_warning is 'A comma separated list of
 comment on column console_conf.units_level_info    is 'A comma separated list of units configured for level info.';
 comment on column console_conf.units_level_debug   is 'A comma separated list of units configured for level debug.';
 comment on column console_conf.units_level_trace   is 'A comma separated list of units configured for level trace.';
+comment on column console_conf.enable_ascii_art    is 'Currently used to have more fun with the APEX error handling messages. But who knows...';
 
 
 
@@ -165,6 +164,7 @@ comment on column console_logs.log_id            is 'Primary key based on a sequ
 comment on column console_logs.log_systime       is 'Log systimestamp.';
 comment on column console_logs.level_id          is 'Level ID. Can be 0 (permanent), 1 (error), 2 (warning), 3 (info), 4 (debug) or 5 (trace).';
 comment on column console_logs.level_name        is 'Level name. Can be Permanent, Error, Warning, Info or Verbose.';
+comment on column console_logs.permanent         is 'If Y the log entry will not be deleted when calling CONSOLE.PURGE or CONSOLE.PURGE_ALL.';
 comment on column console_logs.scope             is 'The current unit/module in which the log was generated (OWNER.PACKAGE.MODULE.SUBMODULE, line number). Couls also be an external scope provided by the user.';
 comment on column console_logs.message           is 'The log message itself.';
 comment on column console_logs.error_code        is 'The error code. Is normally the SQLCODE, but could also be a user error code when log entry was coming from external (user interface, ETL preprocessing, whatever...)';
@@ -246,7 +246,7 @@ prompt - Package CONSOLE (spec)
 create or replace package console authid definer is
 
 c_name    constant varchar2 ( 30 byte ) := 'Oracle Instrumentation Console'       ;
-c_version constant varchar2 ( 20 byte ) := '1.0-beta4'                            ;
+c_version constant varchar2 ( 20 byte ) := '1.0-beta5'                            ;
 c_url     constant varchar2 ( 40 byte ) := 'https://github.com/ogobrecht/console' ;
 c_license constant varchar2 (  5 byte ) := 'MIT'                                  ;
 c_author  constant varchar2 ( 15 byte ) := 'Ottmar Gobrecht'                      ;
@@ -1151,16 +1151,34 @@ $end
 --------------------------------------------------------------------------------
 
 procedure conf (
-  p_level               integer default c_level_error , -- Level 1 (error), 2 (warning), 3 (info), 4 (debug) or 5 (trace).
-  p_check_interval      integer default 10            , -- The number of seconds a session looks for a changed configuration. Allowed values: 1 to 60 seconds.
-  p_units_level_warning varchar2 default null         , -- A comma separated list of units names which should have log level warning. Example: p_units_level_warning => 'OWNER.UNIT,SCHEMA2.PACKAGE3'
-  p_units_level_info    varchar2 default null         , -- Same as p_units_level_warning for level info.
-  p_units_level_debug   varchar2 default null         , -- Same as p_units_level_warning for level debug.
-  p_units_level_trace   varchar2 default null           -- Same as p_units_level_warning for level trace.
+  p_level               integer  default c_level_error , -- Level 1 (error), 2 (warning) or 3 (info).
+  p_check_interval      integer  default 10            , -- The number of seconds a session looks for a changed configuration. Allowed values: 1 to 60 seconds.
+  p_units_level_warning varchar2 default null          , -- A comma separated list of units names which should have log level warning. Example: p_units_level_warning => 'OWNER.UNIT,SCHEMA2.PACKAGE3'
+  p_units_level_info    varchar2 default null          , -- Same as p_units_level_warning for level info.
+  p_units_level_debug   varchar2 default null          , -- Same as p_units_level_warning for level debug.
+  p_units_level_trace   varchar2 default null          , -- Same as p_units_level_warning for level trace.
+  p_enable_ascii_art    boolean  default false           -- Currently used to have more fun with the APEX error handling messages. But who knows...
 );
 /**
 
 Set the global console configuration.
+
+EXAMPLE
+
+```sql
+--set all sessions to level warning
+exec console.conf(p_level => console.c_level_warning);
+
+--set all session to level info and two new packages to debug
+begin
+  console.conf(
+    p_level             => console.c_level_info,
+    p_check_interval    => 10,
+    p_units_level_debug => 'MY_SCHEMA.SOME_API,MY_SCHEMA.ANOTHER_API'
+  );
+end;
+{{/}}
+```
 
 **/
 
@@ -1336,6 +1354,17 @@ Returns `Y` when the input is true and `N` if the input is false or null.
 
 --------------------------------------------------------------------------------
 
+function to_string ( p_bool boolean ) return varchar2;
+/**
+
+Converts a boolean value to a string.
+
+Returns `true` when the input is true and `false` if the input is false or null.
+
+**/
+
+--------------------------------------------------------------------------------
+
 function to_bool ( p_string varchar2 ) return boolean;
 /**
 
@@ -1400,6 +1429,18 @@ begin
 end;
 {{/}}
 ```
+
+**/
+
+--------------------------------------------------------------------------------
+
+function to_md_code_block (
+  p_text  varchar2 )
+return varchar2;
+/**
+
+Converts the given text to a Markdown code block by indent each line with four
+spaces.
 
 **/
 
@@ -1759,8 +1800,6 @@ procedure flush_cache;
 
 Flushes the log cache and writes down the entries to the log table.
 
-Also see clob_append above.
-
 **/
 
 --------------------------------------------------------------------------------
@@ -1775,9 +1814,6 @@ size greater then zero (for example 1000) and you take a look at the log entries
 with the pipelined function `console.view_cache` or
 `console.view_last([numRows])` during development. By clearing the cache you can
 avoid spoiling your CONSOLE_LOGS table with entries you do not need anymore.
-
-DO NOT USE THIS PROCEDURE IN YOUR BUSINESS LOGIC. IT IS INTENDED ONLY FOR
-MANAGING LOGGING MODES OF SESSIONS.
 
 **/
 
@@ -1988,6 +2024,7 @@ g_conf_user_env             boolean;
 g_conf_apex_env             boolean;
 g_conf_cgi_env              boolean;
 g_conf_console_env          boolean;
+g_conf_enable_ascii_art     boolean;
 g_conf_unit_levels          unit_list_tab;
 
 -------------------------------------------------------------------------------
@@ -2733,60 +2770,21 @@ function apex_error_handling (
 return apex_error.t_error_result
 is
   v_result          apex_error.t_error_result;
-  v_reference_id    number;
+  v_log_id          number;
   v_constraint_name t_vc256;
-  v_message         clob;
-  v_app_id          pls_integer;
+  v_app_id          pls_integer := v('APP_ID');
+  v_app_page_id     pls_integer := v('APP_PAGE_ID');
   --
   function extract_constraint_name(p_sqlerrm varchar2) return varchar2 is
   begin
     return regexp_substr(p_sqlerrm, '\(\S+?\.(\S+?)\)', 1, 1, 'i', 1);
   end;
   --
-  procedure create_apex_lang_message ( p_constraint_name varchar2 ) is
-    pragma autonomous_transaction;
-  begin
-    apex_lang.create_message(
-      p_application_id => v_app_id,
-      p_name           => p_constraint_name,
-      p_language       => apex_util.get_preference('FSP_LANGUAGE_PREFERENCE'),
-      p_message_text   => 'DEVELOPER TODO: Change this message for constraint ' || p_constraint_name);
-    commit;
-  end;
-  --
-begin
-  v_app_id := v('APP_ID');
-  v_result := apex_error.init_error_result (p_error => p_error);
-
-  -- If it's an internal error raised by APEX, like an invalid statement or
-  -- code which can't be executed, the error text might contain security sensitive
-  -- information. To avoid this security problem we can rewrite the error to
-  -- a generic error message and log the original error message for further
-  -- investigation by the help desk.
-  if p_error.is_internal_error then
-    -- mask all errors that are not common runtime errors (Access Denied
-    -- errors raised by application / page authorization and all errors
-    -- regarding session and session state)
-    if not p_error.is_common_runtime_error then
-      -- log error for example with an autonomous transaction and return
-      -- v_reference_id as reference#
-      v_message :=
-        case when p_error.message is not null then p_error.message || c_lf end ||
-        case when p_error.additional_info is not null then p_error.additional_info || c_lf end ||
-        case when p_error.error_statement is not null then p_error.error_statement || c_lf end;
-        --FIXME what about other attributes like p_error.component?
-      v_reference_id := error (
-        p_message         => v_message               ,
-        p_call_stack      => false                   ,
-        p_apex_env        => true                    ,
-        p_user_error_code => p_error.ora_sqlcode     ,
-        p_user_call_stack => p_error.error_backtrace );
-      -- Change the message to the generic error message which doesn't expose
-      -- any sensitive information.
-      v_result.message :=
-        $if $$apex_fun $then
-        replace(replace(q'[<pre>
-
+  function get_ascii_art (
+    p_type varchar2 )
+  return varchar2 is
+    v_return varchar2(1000 byte);
+    v_troll  varchar2(1000 byte) := q'[
                 \|||/
                 (o o)
     ,-------ooO--(_)------------.
@@ -2796,70 +2794,183 @@ begin
     '----------------Ooo--------'
                |__|__|
                 || ||
-               ooO Ooo
+               ooO Ooo   ]' || c_lf;
+  begin
+    if g_conf_enable_ascii_art then
+      v_return :=
+        case p_type when 'html' then '<pre>' when 'md' then c_lflf||'```' end             ||
+        replace(replace(v_troll,
+          '#APP_ID##'        , rpad( v_app_id                             ,  9, ' ') ),
+          '#LOG_ID##########', rpad( nvl(to_char(v_log_id), 'this.log_id'), 17, ' ') )  ||
+        case p_type when 'html' then '</pre>' when 'md' then '```' end          ;
+    end if;
+    return v_return;
+  end get_ascii_art;
+  --
+  function create_apex_lang_message (
+    p_constraint_name varchar2
+  ) return varchar2
+  is
+    pragma autonomous_transaction;
+    v_message_text varchar2(4000) :=
+      'DEVELOPER TODO: Change the message in APEX > Application Builder > Shared Components > Text Messages for constraint ' ||
+      p_constraint_name || '.';
+  begin
+    apex_lang.create_message(
+      p_application_id => v_app_id,
+      p_name           => p_constraint_name,
+      p_language       => nvl(apex_util.get_preference('FSP_LANGUAGE_PREFERENCE'), 'en'),
+      p_message_text   => v_message_text);
+    commit;
+    return v_message_text;
+  end create_apex_lang_message;
+  --
+  function get_md_li_pre (p_text varchar2) return varchar2 is
+    v_fences varchar2(30) := '    ```';
+  begin
+    return
+      c_lf ||
+      v_fences || c_lf ||
+      to_md_code_block(p_text) || c_lf ||
+      v_fences;
+  end get_md_li_pre;
+  --
+  function remove_linebreaks (p_text varchar2) return varchar2 is
+  begin
+    return replace(replace(p_text, c_crlf, c_lf), c_lf, ' ');
+  end remove_linebreaks;
+  --
+  function get_log_message (
+    p_text varchar2 )
+  return clob is
+    v_clob  clob;
+    v_cache varchar2(32767);
+  begin
+    clob_append ( v_clob, v_cache, p_text                   || c_lflf                                         );
+    clob_append ( v_clob, v_cache, '## Technical Info'      || c_lflf                                         );
+    clob_append ( v_clob, v_cache, '1. is_internal_error: ' || to_string(p_error.is_internal_error)   || c_lf );
+    clob_append ( v_clob, v_cache, '2. apex_error_code: '   || p_error.apex_error_code                || c_lf );
+    clob_append ( v_clob, v_cache, '3. original message: '  || p_error.message                        || c_lf );
+    clob_append ( v_clob, v_cache, '4. ora_sqlcode: '       || p_error.ora_sqlcode                    || c_lf );
+    clob_append ( v_clob, v_cache, '5. ora_sqlerrm: '       || remove_linebreaks(p_error.ora_sqlerrm) || c_lf );
+    clob_append ( v_clob, v_cache, '6. component.type: '    || p_error.component.type                 || c_lf );
+    clob_append ( v_clob, v_cache, '7. component.id: '      || p_error.component.id                   || c_lf );
+    clob_append ( v_clob, v_cache, '8. component.name: '    || p_error.component.name                 || c_lf );
+    clob_append ( v_clob, v_cache, '9. error_backtrace: '   || get_md_li_pre(p_error.error_backtrace) || c_lf );
+    clob_append ( v_clob, v_cache, '10. error_statement: '  || get_md_li_pre(p_error.error_statement) || c_lf );
+    clob_flush_cache ( v_clob, v_cache );
+    return v_clob;
+  end get_log_message;
+  --
+begin
+  v_result := apex_error.init_error_result (p_error => p_error);
 
-</pre>]', '#APP_ID##', rpad(v_app_id, 9, ' ')),
-          '#LOG_ID##########', rpad(to_char(v_reference_id), 17, ' ')) ||
-        $end
+  -- If it's an internal error raised by APEX, like an invalid statement or code
+  -- which can't be executed, the error text might contain security sensitive
+  -- information. To avoid this security problem we can rewrite the error to a
+  -- generic error message and log the original error message for further
+  -- investigation by the help desk.
+  if p_error.is_internal_error then
+
+    -- Mask all errors that are not common runtime errors (Access Denied errors
+    -- raised by application / page authorization and all errors regarding
+    -- session and session state).
+    if not p_error.is_common_runtime_error then
+
+      -- Log error and return log ID as reference.
+      v_log_id := error (
+        p_message         => get_log_message('Unexpected internal application error.' || get_ascii_art('md')) ,
+        p_call_stack      => false                                                                            ,
+        p_apex_env        => true                                                                             ,
+        p_user_scope      => 'APEX BACKEND ERROR HANDLER: App ' || v_app_id || ', page ' || v_app_page_id     ,
+        p_user_error_code => p_error.ora_sqlcode                                                              ,
+        p_user_call_stack => '## Error Backtrace' || c_lflf || p_error.error_backtrace                        );
+
+      -- Change the message to the generic error message which doesn't expose
+      -- any sensitive information.
+      v_result.message := get_ascii_art('html') ||
         'An unexpected internal application error has occurred. ' ||
         'Please get in contact with your Oracle APEX support team and provide ' ||
-        '"Application ID ' || to_char(v_app_id) || '" and "Log ID ' ||
-         to_char(v_reference_id) || '" for further investigation.';
+        '"App ID ' || to_char(v_app_id) || ', Log ID ' || to_char(v_log_id) ||
+        '" for further investigation.';
       v_result.additional_info := null;
+
     end if;
+
   else
-    -- Always show the error as inline error
-    -- Note: If you have created manual tabular forms (using the package
-    --       apex_item/htmldb_item in the SQL statement) you should still
-    --       use "On error page" on that pages to avoid loosing entered data
+
+    -- Always show the error as inline error.
+    --
+    -- NOTE: If you have created manual tabular forms (using the package
+    --       apex_item/htmldb_item in the SQL statement) you should still use
+    --       "On error page" on that pages to avoid loosing entered data.
     v_result.display_location :=
       case when v_result.display_location = apex_error.c_on_error_page
         then apex_error.c_inline_in_notification
         else v_result.display_location
       end;
 
-    --
-    -- Note: If you want to have friendlier ORA error messages, you can also define
-    --       a text message with the name pattern APEX.ERROR.ORA-number
+    -- NOTE: If you want to have friendlier ORA error messages, you can also
+    --       define a text message with the name pattern APEX.ERROR.ORA-number
     --       There is no need to implement custom code for that.
     --
-
     -- If it's a constraint violation like
     --
-    --   -) ORA-00001: unique constraint violated
-    --   -) ORA-02091: transaction rolled back (-> can hide a deferred constraint)
-    --   -) ORA-02290: check constraint violated
-    --   -) ORA-02291: integrity constraint violated - parent key not found
-    --   -) ORA-02292: integrity constraint violated - child record found
+    -- * ORA-00001: unique constraint violated
+    -- * ORA-02091: transaction rolled back (-> can hide a deferred constraint)
+    -- * ORA-02290: check constraint violated
+    -- * ORA-02291: integrity constraint violated - parent key not found
+    -- * ORA-02292: integrity constraint violated - child record found
     --
-    -- we try to get a friendly error message from our constraint lookup configuration.
-    -- If we don't find the constraint in our lookup table we fallback to
-    -- the original ORA error message.
+    -- We try to get a friendly error message from APEX text messages. If we
+    -- don't find the constraint name there we create a new text message, that
+    -- has to be changed by the developers.
     if p_error.ora_sqlcode in (-1, -2091, -2290, -2291, -2292) then
-      v_constraint_name :=  extract_constraint_name(p_error.ora_sqlerrm);
+      v_constraint_name := extract_constraint_name(p_error.ora_sqlerrm);
       v_result.message := apex_lang.message( v_constraint_name );
+
       if v_result.message = v_constraint_name then
-        --Idea by Roel Hartman: https://roelhartman.blogspot.com/2021/02/stop-using-validations-for-checking.html
-        create_apex_lang_message (v_constraint_name);
+
+        -- * Idea by Roel Hartman:
+        --   https://roelhartman.blogspot.com/2021/02/stop-using-validations-for-checking.html
+        -- * Also see this video by Anton and Neelesh:
+        --   https://www.insum.ca/episode-22-error-handling/
+        v_result.message := create_apex_lang_message (v_constraint_name);
+
+        -- Log a permanent error, so developers get information that they need
+        -- to change the text message.
+        error (
+          p_message         => get_log_message (v_result.message)                                           ,
+          p_permanent       => true                                                                         ,
+          p_call_stack      => false                                                                        ,
+          p_apex_env        => true                                                                         ,
+          p_user_scope      => 'APEX BACKEND ERROR HANDLER: App ' || v_app_id || ', page ' || v_app_page_id ,
+          p_user_error_code => p_error.ora_sqlcode                                                          ,
+          p_user_call_stack => '## Error Backtrace' || c_lflf || p_error.error_backtrace                    );
+
       end if;
+
     end if;
 
-    -- If an ORA error has been raised, for example a raise_application_error(-20xxx, '...')
-    -- in a table trigger or in a PL/SQL package called by a process and we
-    -- haven't found the error in our lookup table, then we just want to see
-    -- the actual error text and not the full error stack with all the ORA error numbers.
+    -- If an ORA error has been raised, for example a
+    -- raise_application_error(-20xxx, '...') in a table trigger or in a PL/SQL
+    -- package called by a process and we haven't found the error in our lookup
+    -- table, then we just want to see the actual error text and not the full
+    -- error stack with all the ORA error numbers.
     if p_error.ora_sqlcode is not null and v_result.message = p_error.message then
       v_result.message := apex_error.get_first_ora_error_text (p_error => p_error);
     end if;
 
     -- If no associated page item/tabular form column has been set, we can use
     -- apex_error.auto_set_associated_item to automatically guess the affected
-    -- error field by examine the ORA error for constraint names or column names.
+    -- error field by examine the ORA error for constraint names or column
+    -- names.
     if v_result.page_item_name is null and v_result.column_alias is null then
       apex_error.auto_set_associated_item (
         p_error        => p_error,
         p_error_result => v_result );
     end if;
+
   end if;
 
   return v_result;
@@ -2973,7 +3084,8 @@ procedure conf (
   p_units_level_warning varchar2 default null          ,
   p_units_level_info    varchar2 default null          ,
   p_units_level_debug   varchar2 default null          ,
-  p_units_level_trace   varchar2 default null          )
+  p_units_level_trace   varchar2 default null          ,
+  p_enable_ascii_art    boolean  default false         )
 is
   pragma autonomous_transaction;
   v_old_conf console_conf%rowtype;
@@ -3034,15 +3146,14 @@ begin
     p_check_interval between 10 and 60,
     'Check interval needs to be between 10 and 60 (seconds). ' ||
     'Values between 1 and 10 seconds can only be set per session with the procedure init.');
-  v_conf.conf_id        := c_conf_id;
-  v_conf.conf_by        := substrb(
-                            coalesce(sys_context('USERENV','OS_USER'), sys_context('USERENV','SESSION_USER')),
-                            1,
-                            64);
-  v_conf.conf_sysdate   := sysdate;
-  v_conf.level_id       := p_level;
-  v_conf.level_name     := get_level_name(p_level);
-  v_conf.check_interval := p_check_interval;
+  v_conf.conf_id          := c_conf_id;
+  v_conf.conf_by          := substrb(coalesce(sys_context('USERENV','OS_USER'), sys_context('USERENV','SESSION_USER')), 1, 64);
+  v_conf.conf_sysdate     := sysdate;
+  v_conf.level_id         := p_level;
+  v_conf.level_name       := get_level_name(p_level);
+  v_conf.check_interval   := p_check_interval;
+  v_conf.enable_ascii_art := to_yn(p_enable_ascii_art);
+  --
   normalize_units_and_levels (p_units_level_warning, 2);
   normalize_units_and_levels (p_units_level_info   , 3);
   normalize_units_and_levels (p_units_level_debug  , 4);
@@ -3060,6 +3171,7 @@ begin
   if nvl(v_old_conf.level_id, 1) != v_conf.level_id then
     utl_ctx_clear_all;
   end if;
+  utl_load_session_configuration;
 end conf;
 
 --------------------------------------------------------------------------------
@@ -3241,6 +3353,15 @@ end;
 
 --------------------------------------------------------------------------------
 
+function to_string (
+  p_bool boolean )
+return varchar2 is
+begin
+  return case when p_bool then 'true' else 'false' end;
+end;
+
+--------------------------------------------------------------------------------
+
 function to_bool (
   p_string varchar2 )
 return boolean is
@@ -3392,6 +3513,20 @@ begin
   close_cursor(v_cursor_id);
   return v_clob;
 end to_html_table;
+
+--------------------------------------------------------------------------------
+
+function to_md_code_block (
+  p_text varchar2 )
+return varchar2 is
+begin
+  return
+    '    ' ||
+    rtrim(
+      trim( replace(replace(p_text, c_crlf, c_lf), c_lf, c_lf||'    ') ),
+      c_lf
+    );
+end;
 
 --------------------------------------------------------------------------------
 
@@ -3762,6 +3897,7 @@ begin
   append_row('g_conf_apex_env',                   to_yn( g_conf_apex_env                               ) );
   append_row('g_conf_cgi_env',                    to_yn( g_conf_cgi_env                                ) );
   append_row('g_conf_console_env',                to_yn( g_conf_console_env                            ) );
+  append_row('g_conf_enable_ascii_art',           to_yn( g_conf_enable_ascii_art                       ) );
   append_row('g_conf_unit_levels(2)',                    g_conf_unit_levels(2)                           );
   append_row('g_conf_unit_levels(3)',                    g_conf_unit_levels(3)                           );
   append_row('g_conf_unit_levels(4)',                    g_conf_unit_levels(4)                           );
@@ -4000,8 +4136,12 @@ end;
 function view_status return tab_key_value pipelined is
   v_row rec_key_value;
 begin
+  if g_conf_check_sysdate < sysdate then
+    utl_load_session_configuration;
+  end if;
   pipe row(new rec_key_value('c_version',                       to_char( c_version                                      )) );
   pipe row(new rec_key_value('g_conf_context_is_available',       to_yn( g_conf_context_is_available                    )) );
+  pipe row(new rec_key_value('c_ctx_namespace',                          c_ctx_namespace                                 ) );
   pipe row(new rec_key_value('g_conf_check_sysdate',            to_char( g_conf_check_sysdate,       c_ctx_date_format  )) );
   pipe row(new rec_key_value('g_conf_exit_sysdate',             to_char( g_conf_exit_sysdate,        c_ctx_date_format  )) );
   pipe row(new rec_key_value('g_conf_client_identifier',                 g_conf_client_identifier                        ) );
@@ -4014,6 +4154,11 @@ begin
   pipe row(new rec_key_value('g_conf_apex_env',                   to_yn( g_conf_apex_env                                )) );
   pipe row(new rec_key_value('g_conf_cgi_env',                    to_yn( g_conf_cgi_env                                 )) );
   pipe row(new rec_key_value('g_conf_console_env',                to_yn( g_conf_console_env                             )) );
+  pipe row(new rec_key_value('g_conf_enable_ascii_art',           to_yn( g_conf_enable_ascii_art                        )) );
+  pipe row(new rec_key_value('g_conf_unit_levels(2)',                    g_conf_unit_levels(2)                           ) );
+  pipe row(new rec_key_value('g_conf_unit_levels(3)',                    g_conf_unit_levels(3)                           ) );
+  pipe row(new rec_key_value('g_conf_unit_levels(4)',                    g_conf_unit_levels(4)                           ) );
+  pipe row(new rec_key_value('g_conf_unit_levels(5)',                    g_conf_unit_levels(5)                           ) );
   pipe row(new rec_key_value('g_counters.count',                to_char( g_counters.count                               )) );
   pipe row(new rec_key_value('g_timers.count',                  to_char( g_timers.count                                 )) );
   pipe row(new rec_key_value('g_log_cache.count',               to_char( g_log_cache.count                              )) );
@@ -4328,6 +4473,16 @@ procedure utl_load_session_configuration is
   v_session_conf console_sessions%rowtype;
   v_global_conf  console_conf%rowtype;
   --
+  procedure load_global_conf is
+  begin
+    v_global_conf := utl_read_global_conf;
+    g_conf_unit_levels(2)   :=           v_global_conf.units_level_warning   ;
+    g_conf_unit_levels(3)   :=           v_global_conf.units_level_info      ;
+    g_conf_unit_levels(4)   :=           v_global_conf.units_level_debug     ;
+    g_conf_unit_levels(5)   :=           v_global_conf.units_level_trace     ;
+    g_conf_enable_ascii_art := to_bool ( v_global_conf.enable_ascii_art    ) ;
+  end load_global_conf;
+  --
   procedure set_default_config is
   begin
     --We have no real conf until now, so we fake 24 hours.
@@ -4368,11 +4523,8 @@ procedure utl_load_session_configuration is
   end load_config_from_table_row;
   --
 begin
-  v_global_conf := utl_read_global_conf;
-  g_conf_unit_levels(2) := v_global_conf.units_level_warning ;
-  g_conf_unit_levels(3) := v_global_conf.units_level_info    ;
-  g_conf_unit_levels(4) := v_global_conf.units_level_debug   ;
-  g_conf_unit_levels(5) := v_global_conf.units_level_trace   ;
+  load_global_conf;
+  --
   if g_conf_context_is_available then
     g_conf_exit_sysdate := to_date(sys_context(c_ctx_namespace, c_ctx_exit_sysdate), c_ctx_date_format);
     if g_conf_exit_sysdate is null then
@@ -4392,6 +4544,7 @@ begin
       load_config_from_table_row;
     end if;
   end if;
+  --
   g_conf_check_sysdate := least(g_conf_exit_sysdate, sysdate + 1/24/60/60 * g_conf_check_interval);
 
 end utl_load_session_configuration;