+++ docs/changelog.md
@@ -1,4 +1,24 @@
+<!-- nav -->
+
+[Index](README.md)
+| [Installation](installation.md)
+| [Getting Started](getting-started.md)
+| [API Overview](api-overview.md)
+| [Package Console](package-console.md)
+| [Changelog](changelog.md)
+| [Uninstallation](uninstallation.md)
+
+<!-- navstop -->
+
 # Changelog
+
+## v1.0-beta5 (2021-05-02)
+
+- Fix table CONSOLE_CONF (remove `organization index`)
+- New parameter for conf method
+- Add missing table comments
+- Improved APEX error handler function
+- Improved docs
 
 ## v1.0-beta4 (2021-04-25)